requirements.txt
discord.py==2.3.2
Flask==3.0.2
schedule==1.2.1
SQLAlchemy==2.0.28
